Rename events package to event
