+++ src/main/kotlin/com/dumbdogdiner/stickysurvival/event/TributeAddEvent.kt
@@ -1,10 +1,10 @@
-package com.dumbdogdiner.stickysurvival.events
+package com.dumbdogdiner.stickysurvival.event
 
 import org.bukkit.entity.Player
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
 
-class TributeRemoveEvent(val player: Player) : Event() {
+class TributeAddEvent(val player: Player) : Event() {
 
     companion object {
         private val HANDLERS = HandlerList()
